refactor: implemented suggested changes

+++ src/openrct2/management/Award.cpp
@@ -293,7 +293,7 @@ static bool award_is_deserved_best_food(int32_t activeAwardTypes)
 
     uint32_t shops = 0;
     uint32_t uniqueShops = 0;
-    uint32_t shopTypes = 0;
+    uint64_t shopTypes = 0;
     for (const auto& ride : GetRideManager())
     {
         if (ride.status != RIDE_STATUS_OPEN)
@@ -337,7 +337,7 @@ static bool award_is_deserved_worst_food(int32_t activeAwardTypes)
 
     uint32_t shops = 0;
     uint32_t uniqueShops = 0;
-    uint32_t shopTypes = 0;
+    uint64_t shopTypes = 0;
     for (const auto& ride : GetRideManager())
     {
         if (ride.status != RIDE_STATUS_OPEN)
